backend/requirements.txt
Django==2.2.19
djangorestframework==3.12.4
Pillow>=8.2.0
